+++ jobs.h
@@ -46,8 +46,8 @@
  */
 
 struct procstat {
-	short pid;		/* process id */
-	short status;		/* status flags (defined above) */
+	int pid;		/* process id */
+	int status;		/* status flags (defined above) */
 	char *cmd;		/* text of command being run */
 };
 
@@ -60,8 +60,8 @@ struct procstat {
 struct job {
 	struct procstat ps0;	/* status of process */
 	struct procstat *ps;	/* status or processes when more than one */
-	short nprocs;		/* number of processes */
-	short pgrp;		/* process group of this job */
+	int nprocs;		/* number of processes */
+	int pgrp;		/* process group of this job */
 	char state;		/* true if job is finished */
 	char used;		/* true if this entry is in used */
 	char changed;		/* true if status has changed */
@@ -70,7 +70,7 @@ struct job {
 #endif
 };
 
-extern short backgndpid;	/* pid of last background process */
+extern int backgndpid;	/* pid of last background process */
 extern int job_warning;		/* user was warned about stopped jobs */
 
 void setjobctl __P((int));
